backend/pyproject.toml
[project]
name = "revenue-intelligence-backend"
version = "0.1.0"
description = "Enterprise Revenue Intelligence Platform — backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pyyaml>=6.0.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "celery[redis]>=5.4.0",
  "openpyxl>=3.1.2",
  "python-multipart>=0.0.12",
  "bcrypt>=4.0.1",
  "pyjwt[crypto]>=2.9.0",
  "email-validator>=2.2.0",
  "cryptography>=42.0.0",
  "authlib>=1.3.2",
  "defusedxml>=0.7.1",
  "signxml>=4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",  # session loop scope; 0.23 causes async engine / loop mismatch
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
"app" = ["core/semantic_layer.yaml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# One event loop for the whole test session (matches process-wide async SQLAlchemy engine pool).
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]
